Update dependencies (including tornado)
The update to the tornado dependency takes it from 6.3.2 to 6.3.3,
patching GHSA-qppv-j76h-2rpx (no known CVE ID).

Among the other dependencies this updates is mypy, which is now
conditionally specified so it can use 1.5.8* on Python >=3.8 while
still using 1.4.* on Python 3.7. (I don't want to drop support for
3.7 from this library just yet, but mypy has dropped support for it
starting in 1.5.0, which is understandable since the PSF no longer
supports 3.7.)